// apps/web/generated/api/model/walletContract.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cashmemo V1 API
 * OpenAPI spec version: 1.0.0
 */
import type { WalletBalanceContract } from './walletBalanceContract';

export interface WalletContract {
  /** @nullable */
  archived_at?: string | null;
  balance: WalletBalanceContract;
  currency: string;
  id: string;
  name: string;
  opening_balance: string;
}